Lab02/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
aiohttp==3.8.1
    # via dataframe-image
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via
    #   aiohttp
    #   jsonschema
beautifulsoup4==4.11.1
    # via
    #   dataframe-image
    #   nbconvert
bleach==5.0.1
    # via nbconvert
certifi==2022.6.15
    # via requests
charset-normalizer==2.0.12
    # via
    #   aiohttp
    #   requests
cycler==0.11.0
    # via matplotlib
dataframe-image==0.1.1
    # via -r requirements.in
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.4
    # via jupyter-client
fastjsonschema==2.16.1
    # via nbformat
fonttools==4.37.1
    # via matplotlib
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.3
    # via
    #   requests
    #   yarl
jinja2==3.1.2
    # via nbconvert
jsonschema==4.15.0
    # via nbformat
jupyter-client==7.3.5
    # via nbclient
jupyter-core==4.11.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.2.2
    # via nbconvert
kiwisolver==1.4.4
    # via matplotlib
lxml==4.9.1
    # via nbconvert
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib==3.5.3
    # via
    #   -r requirements.in
    #   dataframe-image
mistune==2.0.4
    # via nbconvert
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
nbclient==0.6.7
    # via nbconvert
nbconvert==7.0.0
    # via dataframe-image
nbformat==5.4.0
    # via
    #   nbclient
    #   nbconvert
nest-asyncio==1.5.5
    # via
    #   jupyter-client
    #   nbclient
numpy==1.23.2
    # via
    #   matplotlib
    #   pandas
packaging==21.3
    # via
    #   matplotlib
    #   nbconvert
pandas==1.4.3
    # via
    #   -r requirements.in
    #   dataframe-image
pandocfilters==1.5.0
    # via nbconvert
pillow==9.2.0
    # via matplotlib
pygments==2.13.0
    # via nbconvert
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   jupyter-client
    #   matplotlib
    #   pandas
python-dotenv==0.21.0
    # via -r requirements.in
pytz==2022.2.1
    # via pandas
pywin32==304
    # via jupyter-core
pyzmq==23.2.1
    # via jupyter-client
requests==2.27.1
    # via
    #   -r requirements.in
    #   dataframe-image
six==1.16.0
    # via
    #   bleach
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
tinycss2==1.1.1
    # via nbconvert
tornado==6.2
    # via jupyter-client
traitlets==5.3.0
    # via
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
urllib3==1.26.12
    # via requests
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
yarl==1.8.1
    # via aiohttp
